OSS-4 refactor(aws-genai): follow the manage plugin NFS/OFS compatibility pattern
Replace the ./alpha subpath with the single-entry-point mechanism used by
community-plugins workspaces/manage: the new frontend system plugin is the
default export of src/index.ts, while the old frontend system keeps its named
exports from the same entry point, isolated in src/legacy.ts and
src/legacy-plugin.ts and marked @deprecated.

src/plugin.tsx now holds the createFrontendPlugin definition with its
extensions moved to src/extensions/, and the old-system page shell moves to
src/components-ofs/AgentPage so src/components stays shared. Old-system install
instructions move to a dedicated README-OFS.md.
