+++ workspaces/aws/plugins/aws-genai/README.md
@@ -80,12 +80,10 @@ so make sure its stylesheet is imported once in `packages/app/src/index.tsx`:
 import '@backstage/ui/css/styles.css';
 ```
 
-#### New frontend system
-
 Add the plugin to `packages/app/src/App.tsx`:
 
 ```typescript
-import awsGenAiPlugin from '@alithya-oss/backstage-plugins-aws-genai/alpha';
+import awsGenAiPlugin from '@alithya-oss/backstage-plugins-aws-genai';
 
 export default createApp({
   features: [awsGenAiPlugin],
@@ -103,54 +101,18 @@ app:
           path: /assistant/:agentName
 ```
 
-#### Legacy frontend system
-
-Edit `packages/app/src/App.tsx` to add a route for the chat UI page:
-
-```typescript
-import { AgentChatPage } from '@alithya-oss/backstage-plugins-aws-genai';
-
-{
-  /* ... */
-}
-
-const routes = (
-  <FlatRoutes>
-    /* ... */
-    <Route path="/assistant/:agentName" element={<AgentChatPage />} />
-  </FlatRoutes>
-);
-```
-
-Now edit `packages/app/src/components/Root/Root.tsx` to add a menu item:
+The plugin does not contribute a sidebar entry: the route carries an
+`:agentName` parameter, so only your app knows which agent to link to. Add a
+link to a concrete agent path from your app's nav content, for example:
 
 ```tsx
-import { ChatIcon } from '@backstage/core-components';
-
-{
-  /* ... */
-}
-export const Root = ({ children }: PropsWithChildren<{}>) => (
-  <SidebarPage>
-    <Sidebar>
-      {/* ... */}
-      <SidebarGroup label="Menu" icon={<MenuIcon />}>
-        {/* ... */}
-        <SidebarItem
-          icon={ChatIcon}
-          to="assistant/general"
-          text="Chat Assistant"
-        />
-        {/* ... */}
-      </SidebarGroup>
-      {/* ... */}
-    </Sidebar>
-    {/* ... */}
-  </SidebarPage>
-);
+<SidebarItem icon={ChatIcon} to="aws-genai/general" text="Chat Assistant" />
 ```
 
-The URL `assistant/general` means we're going to be using an agent named `general`, which we'll configure below.
+For the old frontend system, see [README-OFS.md](./README-OFS.md).
+
+Note that the old frontend system is deprecated, and support for it will be
+removed from this plugin in a future release.
 
 ### Creating your first agent
