+++ workspaces/aws/plugins/aws-genai/package.json
@@ -17,14 +17,10 @@
   },
   "exports": {
     ".": "./src/index.ts",
-    "./alpha": "./src/alpha.tsx",
     "./package.json": "./package.json"
   },
   "typesVersions": {
     "*": {
-      "alpha": [
-        "src/alpha.tsx"
-      ],
       "package.json": [
         "package.json"
       ]
