+++ workspaces/aws/plugins/aws-genai/src/components-ofs/AgentPage/AgentPage.module.css
@@ -1,0 +1,8 @@
+@layer components {
+  .page {
+    display: flex;
+    flex-direction: column;
+    min-height: 0;
+    height: 100%;
+  }
+}
