+++ workspaces/aws/plugins/aws-genai/src/components/AgentPage/AgentPage.module.css
@@ -1,11 +1,4 @@
 @layer components {
-  .page {
-    display: flex;
-    flex-direction: column;
-    min-height: 0;
-    height: 100%;
-  }
-
   .body {
     display: flex;
     flex: 1 1 auto;
