+++ workspaces/aws/plugins/aws-genai/src/components/AgentPage/index.ts
@@ -14,5 +14,4 @@
  * limitations under the License.
  */
 
-export { AgentPage, AgentPageContent, NfsAgentPage } from './AgentPage';
-export type { AgentPageProps } from './AgentPage';
+export { AgentPageContent } from './AgentPage';
