+++ workspaces/aws/plugins/aws-genai/src/components/index.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export { AgentPage } from './AgentPage';
+export { AgentPageContent } from './AgentPage';
